extract log configuration (#201)
